LayoutTests/imported/w3c/web-platform-tests/fetch/api/basic/request-forbidden-headers.any.worker-expected.txt


PASS Accept-Charset is a forbidden request header
PASS Accept-Encoding is a forbidden request header
PASS Access-Control-Request-Headers is a forbidden request header
PASS Access-Control-Request-Method is a forbidden request header
PASS Connection is a forbidden request header
PASS Content-Length is a forbidden request header
PASS Cookie is a forbidden request header
PASS Cookie2 is a forbidden request header
PASS Date is a forbidden request header
PASS DNT is a forbidden request header
PASS Expect is a forbidden request header
PASS Host is a forbidden request header
PASS Keep-Alive is a forbidden request header
PASS Origin is a forbidden request header
PASS Referer is a forbidden request header
PASS TE is a forbidden request header
PASS Trailer is a forbidden request header
PASS Transfer-Encoding is a forbidden request header
PASS Upgrade is a forbidden request header
PASS Via is a forbidden request header
PASS Proxy- is a forbidden request header
PASS Proxy-Test is a forbidden request header
PASS Sec- is a forbidden request header
PASS Sec-Test is a forbidden request header
